pyproject.toml
[project]
name = "week12-medical-telegram-warehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dagster>=1.6,<1.8",
    "dagster-webserver>=1.6,<1.8",
    "dbt-postgres",
    "fastapi>=0.110,<1.0",
    "pandas>=2.2,<2.3",
    "plotly>=5.18.0",
    "psycopg2-binary>=2.9,<3.0",
    "pydantic>=2.5,<2.7",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0,<2.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0,<2.1",
    "streamlit>=1.31,<1.36",
    "telethon>=1.33,<2.0",
    "torch>=2.10.0",
    "tqdm>=4.66,<5.0",
    "ultralytics>=8.3,<9.0",
    "uvicorn[standard]>=0.29,<0.30",
    "yolov8>=0.0.2",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.0",
    "types-psycopg2>=2.9.21.20251012",
    "types-requests>=2.32.4.20260107",
]

[tool.uv]
environments = ["python_full_version >= '3.12' and sys_platform == 'linux'"]
